# Use a faster deepcopy library