--- a/go.mod
+++ b/go.mod
@@ -3,9 +3,9 @@ module github.com/xuri/excelize/v2
 go 1.18
 
 require (
-	github.com/mohae/deepcopy v0.0.0-20170929034955-c48cc78d4826
 	github.com/richardlehane/mscfb v1.0.4
-	github.com/stretchr/testify v1.8.4
+	github.com/stretchr/testify v1.9.0
+	github.com/tiendc/go-deepcopy v1.1.0
 	github.com/xuri/efp v0.0.0-20240408161823-9ad904a10d6d
 	github.com/xuri/nfp v0.0.0-20240318013403-ab9948c2c4a7
 	golang.org/x/crypto v0.29.0