# PR Details

## Description

<!--- Describe your changes in detail -->

## Related Issue

Closes https://github.com/qax-os/excelize/issues/2029.

## Motivation and Context

Benchmark:

```go
package excelize

import (
	"testing"

	olddeepcopy "github.com/mohae/deepcopy"
	newdeepcopy "github.com/tiendc/go-deepcopy"
)

func BenchmarkCopyXlsxCol(b *testing.B) {
	src := xlsxCol{}

	b.Run("old", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			olddeepcopy.Copy(src)
		}
	})

	b.Run("new", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			var dst xlsxCol
			newdeepcopy.Copy(&dst, src)
		}
	})
}

func BenchmarkCopyXlsxRow(b *testing.B) {
	src := xlsxRow{}

	b.Run("old", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			olddeepcopy.Copy(src)
		}
	})

	b.Run("new", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			var dst xlsxRow
			newdeepcopy.Copy(&dst, src)
		}
	})
}

func BenchmarkCopyXlsxWorksheet(b *testing.B) {
	f, err := prepareTestBook1()
	if err != nil {
		b.Fatal(err)
	}

	idx, err := f.NewSheet("CopySheet")
	if err != nil {
		b.Fatal(err)
	}

	fromSheet := f.GetSheetName(idx)
	src, err := f.workSheetReader(fromSheet)
	if err != nil {
		b.Fatal(err)
	}

	b.Run("old", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			olddeepcopy.Copy(src)
		}
	})

	b.Run("new", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			var dst xlsxWorksheet
			newdeepcopy.Copy(&dst, src)
		}
	})
}
```

Result:

```
goos: linux
goarch: amd64
pkg: github.com/xuri/excelize/v2
cpu: AMD Ryzen 7 PRO 4750U with Radeon Graphics
BenchmarkCopyXlsxCol/old-16      	  446193	      2483 ns/op	     224 B/op	      13 allocs/op
BenchmarkCopyXlsxCol/new-16      	  963056	      1148 ns/op	     128 B/op	       3 allocs/op
BenchmarkCopyXlsxRow/old-16      	  344492	      3135 ns/op	     344 B/op	      16 allocs/op
BenchmarkCopyXlsxRow/new-16      	  909142	      1371 ns/op	     192 B/op	       3 allocs/op
BenchmarkCopyXlsxWorksheet/old-16         	   49108	     23097 ns/op	    3656 B/op	     128 allocs/op
BenchmarkCopyXlsxWorksheet/new-16         	  176534	      7085 ns/op	     808 B/op	       6 allocs/op
PASS
ok  	github.com/xuri/excelize/v2	7.377s
```

## How Has This Been Tested

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- See how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [x] Docs change / refactoring / dependency upgrade
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.