@xuri 2024-11-28
Thanks for your comments, although the new deepcopy library not popular and not widely used currently, but after I checked the code of that, it seems fine. I think if bug reports before next release of excelize, we can revert this or using encoding/gob for deepcopy. @Juneezee what you think of this?