@Juneezee 2024-11-28
Hi @arp242, thank you for your feedback and concerns. I did review the source code of the suggested library before making the proposal in https://github.com/qax-os/excelize/issues/2029.

> The previous library was unmaintained, but it was simple, worked, and was trusted. This is a random project from more or less a random person without any reputation that I can see.

You’re correct that the previous library code is simpler. However, the new library also includes unit tests to ensure functionality. Are your main concerns related to the security or trustworthiness of the code?

> The benchmarks didn't really demonstrate this makes a meaningful difference.

The performance improvement is most noticeable in the Copy Worksheet function. For the benchmark, I used a small worksheet from the test Excel file (containing only two charts and a few rows and columns with data). Even with this relatively simple example, we see significant improvements in execution time and memory allocations. These benefits will scale and become even more impactful when working with larger worksheets.