@shcabin 2025-04-24
> 有人审计过这个图书馆吗？
> 
> 之前的库无人维护，但它简单、可用且值得信赖。这是一个随机项目，几乎出自一个毫无名气的随机开发者之手。虽然我 99% 确定它没问题，但并非 100%。我不太愿意把它引入到我们的依赖关系树中。
> 
> 我们这里讨论的是纳秒级。基准测试并没有真正证明这会带来什么显著的差异。


I have identified a potential issue where this library cannot handle DeepCopy of sync.Map, will return an ErrTypeNonCopyable error. And the code here did not handle any errors