@Juneezee 2025-04-24
> I have identified a potential issue where this library cannot handle DeepCopy of sync.Map, will return an ErrTypeNonCopyable error. And the code here did not handle any errors

@shcabin Thanks for reporting the issue. I have reported the issue to library author here: https://github.com/tiendc/go-deepcopy/issues/34.